Limpeza

+++ app/templates/Competicions/index.php
@@ -2,9 +2,11 @@
 $this->extend('template');
 $this->set('cabeceiraTitulo', 'Competicións');
 $this->set('cabeceiraMigas', [
-    ['label'=>'Competicións'],
-    ['label'=>'Administrar']
+    ['label'=>'Competicións']
 ]);
+
+$authUser = $this->request->getAttribute('identity');
+$is_admin = $authUser['rol'] === 'admin';
 ?>
 
 <div class="row">
@@ -17,14 +19,19 @@ $this->set('cabeceiraMigas', [
                     <th>Tipo</th>
                     <th>Código</th>
                     <th class="column-button"></th>
+                    <?php if($is_admin) : ?>
+                        <th class="column-button"></th>
+                        <th class="column-button"></th>
+                    <?php endif ?>
                 </tr>
             </thead>
             <tbody>
                 <?php $tempada = ""; ?>
+                <?php $colspan = $is_admin ? "7" : "5"; ?>
                 <?php foreach($competicions as $c) : ?>
 
                     <?php if ($tempada !== $c->tempada) : ?>
-                        <th colspan="6" style="text-align: center; line-height: 30px; background-color: #eee">
+                        <th colspan="<?= $colspan ?>" style="text-align: center; line-height: 30px; background-color: #eee">
                             <?= $tempadas[$c->tempada] ?>
                         </th>
                         <?php $tempada = $c->tempada; ?>
@@ -32,10 +39,14 @@ $this->set('cabeceiraMigas', [
                     
                     <tr>
                         <td><?= $this->AgfgForm->logo($c->federacion) ?></td>
-                        <td><?= $this->Html->link($c->nome, ['action'=>'detalle', $c->id]) ?></td>
+                        <td><?= $c->nome ?></td>
                         <td><?= empty($c->tipo) ? '' : $tiposCompeticion[$c->tipo] ?></td>
                         <td><?= $c->codigo ?></td>
-                        <td class="text-center"><?= $this->AgfgForm->deleteButton(['action'=>'borrar', $c->id]) ?></td>
+                        <td><?= $this->Html->link('', ['controller'=>'Resultados', 'action'=>'competicion', $c->id], ['class'=>'glyphicon glyphicon-flag']) ?></td>
+                        <?php if($is_admin) : ?>
+                            <td><?= $this->Html->link('', ['action'=>'detalle', $c->id], ['class'=>'glyphicon glyphicon-cog']) ?></td>
+                            <td class="text-center"><?= $this->AgfgForm->deleteButton(['action'=>'borrar', $c->id]) ?></td>
+                        <?php endif ?>
                     </tr>
                 <?php endforeach ?>
             </tbody>
